Fix add company verification & add company name for edit
